requirements.txt
numpy>=1.24
opencv-python>=4.7
scipy>=1.10
matplotlib>=3.7
Pillow>=10.0
